requirements.txt
# Core Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# OpenAI APIs (GPT-4, Whisper, Embeddings)
openai==1.12.0

# Database
Flask-SQLAlchemy==3.1.1

# File Processing
PyPDF2==3.0.1
Pillow==10.2.0

# QR Code Generation
qrcode[pil]==7.4.2

# Security & Auth
werkzeug==3.0.1
python-dotenv==1.0.0

# Deployment
gunicorn==21.2.0

# Date/Time
python-dateutil==2.8.2

# JSON Processing
jsonschema==4.21.1

# Numerical Computing (for embeddings cosine similarity)
numpy==1.24.3

# HTTP Requests (if needed for API calls)
requests==2.31.0